feat(providers): ask the account, not a table, whether a credential works
`lev setup`'s check called `list_models`, which for every key-based provider
is an authenticated call: a listing that succeeded is a credential that
works, two answers for one round trip. Codex breaks that, because its
catalogue is a table compiled into the build. Its check therefore reported
"5 models" for a revoked session, an expired one, and for never having
signed in at all - the wizard telling you that you are set up when you are
not, which is worse than not checking.

So the check is its own trait method, `check_credential`, defaulting to
exactly the listing every other provider already made. Codex overrides it
with the quota route: authenticated, costs no model tokens, and it comes
back with the plan tier, which is also what decides which of those models
the account may use. A successful check therefore teaches `served_catalog`
the tier, and refreshes a lapsed token on its way in, so checking a
rarely-used sign-in is also what keeps it alive.

The quota route now reads its status before parsing the body. A refused
session answers with a perfectly well-formed error document, so parsing
first turned "your session was revoked" into "not in a known shape".

`provider.rs` was four lines under the production-line limit, so the
shared-helper block below the trait moves to `provider/helpers.rs` and is
re-exported. The seam was already drawn with a banner; nothing's path
changes.
